Split out helix_cli into a new gem
