+++ cli/Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+
+# Specify your gem's dependencies in ruby.gemspec
+gemspec
+
+gem "helix_runtime", path: "../ruby"
