+++ cli/lib/helix_cli/cli/templates/Cargo.toml
