+++ cli/lib/helix_cli/cli/templates/Gemfile
