+++ cli/lib/helix_cli/cli/templates/gitignore
