+++ cli/lib/helix_cli/cli/templates/helix_cli.rake
@@ -1,0 +1,5 @@
+require 'helix_cli/build_task'
+
+HelixCLI::BuildTask.new
+
+task :default => :build
