+++ cli/lib/helix_cli/cli/templates/lib.rs
