+++ cli/lib/helix_cli/parent_project.rb
@@ -1,4 +1,4 @@
-module HelixRuntime
+module HelixCLI
   class ParentProject
 
     attr_accessor :root
