+++ cli/lib/helix_cli/platform.rb
@@ -1,6 +1,6 @@
 require 'rbconfig'
 
-module HelixRuntime
+module HelixCLI
   module Platform
     # Normalize the platform OS
     OS = case os = RbConfig::CONFIG['host_os'].downcase
