--- ruby/lib/helix_runtime/cli/templates/helix_runtime.rake
@@ -1,5 +1,0 @@
-require 'helix_runtime/build_task'
-
-HelixRuntime::BuildTask.new
-
-task :default => :build
